Updated scaffolding
