+++ frontend/src/pages/admin/components/index.js
@@ -1,0 +1,3 @@
+export { PostEdit, PostTitle } from "./posts/EditProduct";
+export { PostCreate, PostIcon, PostList } from "./posts/CreateProduct";
+export { PostList } from "./get/ReadProducts";
