+++ frontend/src/pages/index.js
@@ -1,3 +1,3 @@
 export { Home } from "./home/Home";
 export { AdminPage  } from "./admin/Admin";
-export {PostList, PostEdit, PostCreate, PostIcon} from "./admin/components/posts/ProductCreate";
+export {PostList, PostEdit, PostCreate, PostIcon} from "./admin/components/posts/CreateProduct";
